feat(week-4/amm): implement withdraw instruction
- burn LP tokens before transferring to prevent state inconsistency
- calculate proportional X and Y return using calculate_withdraw
- withdraw_token helper handles both X and Y vault transfers
- pool PDA signs vault transfers
- slippage guard on both min_x and min_y
- no locked check on withdraw — users must always be able to exit
- emit WithdrawEvent with lp_burned, amount_x, amount_y

+++ week-4/amm/programs/amm/src/state/pool.rs
@@ -5,7 +5,7 @@ use anchor_lang::prelude::*;
 pub struct AmmConfig {
     pub fee_rate: u16,             // basis points e.g. 30 = 0.3%
     pub authority: Option<Pubkey>, // who can update this config
-    pub index: u16,                // unique index, used as PDA seed
+    pub index: u64,                // unique index, used as PDA seed
     pub bump: u8,                  // PDA bump saved so we don't recompute
 }
